% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter-biblias.R
\name{baixar_biblias}
\alias{baixar_biblias}
\title{Faz o download das biblias em formato SQLite neste repositório}
\usage{
baixar_biblias()
}
\description{
Faz o download das biblias em formato SQLite neste repositório
}
