% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter-biblias.R
\name{biblia_sql2json}
\alias{biblia_sql2json}
\title{Converte uma versao da biblia no formato SQLite para JSON}
\usage{
biblia_sql2json(versao)
}
\arguments{
\item{versao}{versao da traducao biblica: ACF, ARA, ARC, AS21, JFAA,
KJA, KJF, NAA, NBV, NTLH, NVI, NVT ou TB}
}
\description{
Converte uma versao da biblia no formato SQLite para JSON
}
