% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter-biblias.R
\name{conversao_biblias}
\alias{conversao_biblias}
\title{Converte todas as versoes das biblias no formato SQLite para JSON}
\usage{
conversao_biblias()
}
\description{
Converte todas as versoes das biblias no formato SQLite para JSON
}
