% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigir-template.R
\name{corrigir_capitulo_usx}
\alias{corrigir_capitulo_usx}
\title{Corrige um determinado capitulo de um determinado livro, atualizando
o numero de versiculos, retirando ou adicionando versos.}
\usage{
corrigir_capitulo_usx(livro, capitulo, ver_atual, ver_novo)
}
\arguments{
\item{livro}{inteiro correspondente a ordem do livro da biblia}

\item{capitulo}{inteiro correspondente ao capitulo da biblia}

\item{ver_atual}{numero de versiculos no arquivo usx, a ser substituido}

\item{ver_novo}{numero de versiculos para substituir no arquivo usx}
}
\value{
uma \code{lista}
}
\description{
Corrige um determinado capitulo de um determinado livro, atualizando
o numero de versiculos, retirando ou adicionando versos.
}
