% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigir-template.R
\name{corrigir_usx}
\alias{corrigir_usx}
\title{Faz a correcao de uma versao da biblia, verificando e atualizando
todos os capitulos, adicionando ou removendo versiculos no template
para adequar a versao}
\usage{
corrigir_usx(versao)
}
\arguments{
\item{versao}{versao da traducao biblica: ACF, ARA, ARC, AS21, JFAA,
KJA, KJF, NAA, NBV, NTLH, NVI, NVT ou TB}
}
\description{
Faz a correcao de uma versao da biblia, verificando e atualizando
todos os capitulos, adicionando ou removendo versiculos no template
para adequar a versao
}
