% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduzir-biblias.R
\name{gerar_id_biblia}
\alias{gerar_id_biblia}
\title{Gera um id aleatorio para uma versao qualquer da biblia
para ser colocada no arquivo de metadados da versao}
\usage{
gerar_id_biblia()
}
\description{
Gera um id aleatorio para uma versao qualquer da biblia
para ser colocada no arquivo de metadados da versao
}
