% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter-biblias.R
\name{ler_biblia}
\alias{ler_biblia}
\title{Leitura da Biblia em formato JSON como uma lista aninhada}
\usage{
ler_biblia(versao)
}
\arguments{
\item{versao}{versao da traducao biblica: ACF, ARA, ARC, AS21, JFAA,
KJA, KJF, NBV, NTLH, NVI, NVT ou TB}
}
\value{
uma \code{lista}
}
\description{
Leitura da Biblia em formato JSON como uma lista aninhada
}
