% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigir-template.R
\name{necessario_atualizar}
\alias{necessario_atualizar}
\title{Verifica se e necessario fazer atualizacoes no template para
adequar versoes com diferentes numeros de versiculos}
\usage{
necessario_atualizar(versao)
}
\arguments{
\item{versao}{versao da traducao biblica: ACF, ARA, ARC, AS21, JFAA,
KJA, KJF, NAA, NBV, NTLH, NVI, NVT ou TB}
}
\value{
uma \code{tibble}
}
\description{
Verifica se e necessario fazer atualizacoes no template para
adequar versoes com diferentes numeros de versiculos
}
