% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigir-template.R
\name{numero_versiculos}
\alias{numero_versiculos}
\title{Detalha os versiculos e capitulos de um determinado livro
definido pelo arquivo .usx}
\usage{
numero_versiculos(usx)
}
\arguments{
\item{usx}{objeto usx lido pelo pacote xml2}
}
\value{
uma \code{tibble}
}
\description{
Detalha os versiculos e capitulos de um determinado livro
definido pelo arquivo .usx
}
