% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduzir-biblias.R
\name{traduzir_biblias}
\alias{traduzir_biblias}
\title{Faz a traducao (conversao) dos versos contidos em um JSON
para o formato .usx dado pelo template}
\usage{
traduzir_biblias()
}
\description{
Faz a traducao (conversao) dos versos contidos em um JSON
para o formato .usx dado pelo template
}
