% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduzir-biblias.R
\name{traduzir_livro}
\alias{traduzir_livro}
\title{Faz a traducao (conversao) dos versos contidos em um
livro especifico em JSON para o formato .usx
dado pelo template}
\usage{
traduzir_livro(traduzir, tradutor, traducao)
}
\arguments{
\item{traduzir}{objeto usx lido pelo pacote xml2}

\item{tradutor}{lista contendo o livro da biblia,
lido pelo pacote jsonlite}

\item{traducao}{sigla da versao da biblia utilizada}
}
\value{
uma \code{lista}
}
\description{
Faz a traducao (conversao) dos versos contidos em um
livro especifico em JSON para o formato .usx
dado pelo template
}
