% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduzir-biblias.R
\name{traduzir_metadata}
\alias{traduzir_metadata}
\title{Faz a traducao do arquivo metadata com informacoes da versao
da biblia utilizada}
\usage{
traduzir_metadata()
}
\value{
uma \code{lista}
}
\description{
Faz a traducao do arquivo metadata com informacoes da versao
da biblia utilizada
}
