% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduzir-biblias.R
\name{traduzir_rvmetadata}
\alias{traduzir_rvmetadata}
\title{Faz a traducao do arquivo rvmetadata com informacoes da versao
da biblia utilizada}
\usage{
traduzir_rvmetadata(nome, abbr, abbr_original)
}
\arguments{
\item{nome}{nome completo da versao da biblia utilizada}

\item{abbr}{abreviatura da versao da biblia utilizada}

\item{abbr_original}{abreviatura da versao da biblia do propresenter
utilizada como template}
}
\value{
uma \code{lista}
}
\description{
Faz a traducao do arquivo rvmetadata com informacoes da versao
da biblia utilizada
}
