% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traduzir-biblias.R
\name{zipar_biblias}
\alias{zipar_biblias}
\title{Faz a compressao em .zip de todos os arquivos .usx
de todas biblias}
\usage{
zipar_biblias()
}
\description{
Faz a compressao em .zip de todos os arquivos .usx
de todas biblias
}
